moved cleaner function to util/interrupt
