+++ commands/bridge_pull.go
@@ -4,7 +4,7 @@ import (
 	"github.com/MichaelMure/git-bug/bridge"
 	"github.com/MichaelMure/git-bug/bridge/core"
 	"github.com/MichaelMure/git-bug/cache"
-	"github.com/MichaelMure/git-bug/cleaner"
+	"github.com/MichaelMure/git-bug/util/interrupt"
 	"github.com/spf13/cobra"
 )
 
@@ -14,7 +14,7 @@ func runBridgePull(cmd *cobra.Command, args []string) error {
 		return err
 	}
 	defer backend.Close()
-	cleaner.Register(backend.Close)
+	interrupt.RegisterCleaner(backend.Close)
 
 	var b *core.Bridge
